requirements.txt
flask
line-bot-sdk
redis